refactor segment trajectory into multiple methods
