+++ tests/approaches/test_grammar_search_invention_approach.py
@@ -13,7 +13,7 @@ from predicators.src.predicate_search_score_functions import \
 from predicators.src.envs import CoverEnv
 from predicators.src.structs import Type, Predicate, STRIPSOperator, State, \
     Action, Box, LowLevelTrajectory, Dataset
-from predicators.src.nsrt_learning.strips_learning import segment_trajectory
+from predicators.src.nsrt_learning.segmentation import segment_trajectory
 from predicators.src.settings import CFG
 from predicators.src import utils
